Dimensions

Documentation for Dimensions.

Dimensions.dimensions

Dimensions.find_middle

Dimensions.find_middle

Dimensions.find_midpoints

Dimensions.find_spacing

Dimensions.dimensions — Method

dimensions(
xs::Vector{Vector{T}}, 
ys::Vector{Vector{S}}; 
offset = zero(S), # offset in y direction
) where T where S

Finds the dimensions of a horizontally spaced set of objects with points xs and ys. Each vector in xs and ys represents a new object.

Dimensions.find_middle — Method

find_middle(xs::AbstractArray{T}) where T

Compute the middle of an array xs for each column, which consists of finding its extrema and then computing their mean.

Example

julia> a = [
       1 2 3
       4 5 6
       7 8 9
       ]
3×3 Matrix{Int64}:
 1  2  3
 4  5  6
 7  8  9

julia> find_middle(a)
3-element Vector{Int64}:
 4
 5
 6

Dimensions.find_middle — Method

find_middle(xs::AbstractArray{T}, ys::AbstractArray{S}) where T where S

Compute the middle of an array xs and ys for each column, which consists of finding its extrema and then computing their mean. Will return two vectors for each array.

Dimensions.find_midpoints — Method

find_midpoints(xs::AbstractVector{T}) where T

Compute the midpoints of a vector xs, which consists of computing the average between adjacent values in the vector.

Example

julia> a = [10; 20; 30; 40; 50]
5-element Vector{Int64}:
 10
 20
 30
 40
 50

julia> find_midpoints(a)
4-element Vector{Int64}:
 15
 25
 35
 45

Dimensions.find_spacing — Method

find_spacing(xs::AbstractVector{T}) where T

Compute the spacing of a vector xs, which consists of computing the difference between adjacent values in the vector.

Example

julia> a = [1; 3; 7; 20; 30]
5-element Vector{Int64}:
  1
  3
  7
 20
 30

julia> find_spacing(a)
4-element Vector{Int64}:
  2
  4
 13
 10